+++ lua/plugins/config/theme.lua
@@ -3,14 +3,15 @@ return function(use)
 	_G.loaded_theme = ''
 	_G.load_random_theme = function()
 		local theme_name = _G.theme_list[math.random(#_G.theme_list)]
-		if theme_name == _G.loaded_theme and #_G.theme_list >= 2 then
-			_G.load_random_theme()
-			return
-		end
 		_G.loaded_theme = theme_name
 		vim.cmd('colo ' .. theme_name)
 	end
-	vim.cmd([[au VimEnter * lua _G.load_random_theme()]])
+	_G.load_random_theme_delayed = function()
+		vim.defer_fn(function()
+			_G.load_random_theme()
+		end, 100)
+	end
+	vim.cmd([[au DirChanged * lua _G.load_random_theme_delayed()]])
 	vim.cmd([[command! RandomColor lua _G.load_random_theme()]])
 
 	use({
@@ -24,6 +25,13 @@ return function(use)
 		setup = function()
 			vim.g.tokyonight_italic_functions = true
 			table.insert(_G.theme_list, 'tokyonight')
+		end,
+		config = function()
+			-- draw twice to avoid black bars on startup
+			vim.cmd([[colo tokyonight]])
+			vim.defer_fn(function()
+				vim.cmd([[colo tokyonight]])
+			end, 100)
 		end,
 	})