+++ lua/plugins/config/flutter.lua
@@ -7,7 +7,7 @@ return function(use)
 			require('flutter-tools').setup({
 				debugger = {
 					enabled = true,
-					run_via_dap = true,
+					run_via_dap = false,
 				},
 				widget_guides = {
 					enabled = true,