+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Feat
 - **debugger:** now uses personal one with wrap set to default on
 
+### Fix
+- **theme:** theme now fixed
+
 ### Refactor
 - **golang-lsp:** staticcheck now depends on golangci lint